Updated the imports and some TS to work with the refactored uikit
The refactored UIKit can be found here: https://github.com/AgoraIO-Community/ReactNative-UIKit/pull/39
