+++ template/src/components/Controls.tsx
@@ -17,12 +17,12 @@ import {
   StyleSheet,
   Text,
 } from 'react-native';
-import LocalUserContext from '../../agora-rn-uikit/src/LocalUserContext';
+import {LocalUserContext} from '../../agora-rn-uikit';
 import {
   LocalAudioMute,
   LocalVideoMute,
   Endcall,
-} from '../../agora-rn-uikit/Components';
+} from '../../agora-rn-uikit';
 import Recording from '../subComponents/Recording';
 import SwitchCamera from '../subComponents/SwitchCamera';
 // import icons from '../assets/icons';
