+++ template/src/subComponents/LocalVideoMute.tsx
@@ -1,17 +1,17 @@
 /*
 ********************************************
  Copyright © 2021 Agora Lab, Inc., all rights reserved.
- AppBuilder and all associated components, source code, APIs, services, and documentation 
- (the “Materials”) are owned by Agora Lab, Inc. and its licensors. The Materials may not be 
- accessed, used, modified, or distributed for any purpose without a license from Agora Lab, Inc.  
- Use without a license or in violation of any license terms and conditions (including use for 
- any purpose competitive to Agora Lab, Inc.’s business) is strictly prohibited. For more 
- information visit https://appbuilder.agora.io. 
+ AppBuilder and all associated components, source code, APIs, services, and documentation
+ (the “Materials”) are owned by Agora Lab, Inc. and its licensors. The Materials may not be
+ accessed, used, modified, or distributed for any purpose without a license from Agora Lab, Inc.
+ Use without a license or in violation of any license terms and conditions (including use for
+ any purpose competitive to Agora Lab, Inc.’s business) is strictly prohibited. For more
+ information visit https://appbuilder.agora.io.
 *********************************************
 */
 import React, {useContext} from 'react';
-import RtcContext, {DispatchType} from '../../agora-rn-uikit/src/RtcContext';
-import {LocalContext} from '../../agora-rn-uikit/src/LocalUserContext';
+import {RtcContext} from '../../agora-rn-uikit';
+import {LocalContext} from '../../agora-rn-uikit';
 import {Image, StyleSheet, TouchableOpacity} from 'react-native';
 import icons from '../assets/icons';
 import ColorContext from '../components/ColorContext';
@@ -27,7 +27,7 @@ function LocalVideoMute() {
   return (
     <TouchableOpacity
       onPress={() => {
-        (dispatch as DispatchType<'LocalMuteVideo'>)({
+        dispatch({
           type: 'LocalMuteVideo',
           value: [local.video],
         });
